+++ customer-service/src/main/java/com/microservices/customer/dao/CustomerDAO.java
@@ -1,0 +1,9 @@
+package com.microservices.customer.dao;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.microservices.customer.entity.Customer;
+
+public interface CustomerDAO extends JpaRepository<Customer, Long> {
+
+}
