+++ customer-service/src/main/java/com/microservices/customer/dao/ProductDAo.java
@@ -1,8 +1,8 @@
-package com.microservices.customer.wishlist.dao;
+package com.microservices.customer.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.microservices.customer.wishlist.entity.Product;
+import com.microservices.customer.entity.Product;
 
 public interface ProductDAo extends JpaRepository<Product, Long> {
