+++ customer-service/src/main/java/com/microservices/customer/entity/WishList.java
@@ -1,4 +1,4 @@
-package com.microservices.customer.wishlist.entity;
+package com.microservices.customer.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
