+++ customer-service/src/main/java/com/microservices/customer/service/CustomerService.java
@@ -1,0 +1,15 @@
+package com.microservices.customer.service;
+
+import java.util.List;
+
+import com.microservices.customer.entity.Product;
+
+public interface CustomerService {
+
+	List<Product> getItems(long userId);
+
+	void addWishlistItem(long userId, long productId);
+
+	void removeWishlistItem(long userId, long productId);
+
+}
