+++ customer-service/src/main/java/com/microservices/customer/service/ProductReviewService.java
@@ -1,13 +1,10 @@
-package com.microservices.customer.wishlist.service;
+package com.microservices.customer.service;
 
 import java.util.List;
 
-import com.microservices.customer.wishlist.entity.ProductReview;
-
+import com.microservices.customer.entity.ProductReview;
 
 public interface ProductReviewService {
-
-	
 
 	List<ProductReview> getReviewsByCustomerId(long userId);
