+++ customer-service/src/main/java/com/microservices/customer/service/impl/CustomerServiceImpl.java
@@ -1,0 +1,59 @@
+package com.microservices.customer.service.impl;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import com.microservices.customer.dao.CustomerDAO;
+import com.microservices.customer.dao.ProductDAo;
+import com.microservices.customer.dao.WishlistDAO;
+import com.microservices.customer.entity.Customer;
+import com.microservices.customer.entity.Product;
+import com.microservices.customer.entity.WishList;
+import com.microservices.customer.service.CustomerService;
+
+import jakarta.transaction.Transactional;
+
+@Service
+public class CustomerServiceImpl implements CustomerService {
+
+	@Autowired
+	private WishlistDAO dao;
+
+	@Autowired
+	private CustomerDAO customerDAO;
+
+	@Autowired
+	private ProductDAo productDAO;
+
+	@Override
+	@Transactional
+	public List<Product> getItems(long userId) {
+		List<WishList> wishLists = dao.findByCustomerUserId(userId);
+		return wishLists.stream().map(wl -> productDAO.findById(wl.getProductId()).orElse(null))
+				.collect(Collectors.toList());
+	}
+
+	@Override
+	@Transactional // Ensure transaction for adding an item
+	public void addWishlistItem(long userId, long productId) {
+		Customer customer = customerDAO.findById(userId).orElseThrow(() -> new RuntimeException("Customer not found"));
+
+		WishList wishList = new WishList();
+		wishList.setCustomer(customer);
+		wishList.setProductId(productId);
+
+		dao.save(wishList); // Save to DB
+	}
+
+	@Override
+	@Transactional // Ensure transaction for removing an item
+	public void removeWishlistItem(long userId, long productId) {
+		WishList wishList = dao.findByCustomerUserId(userId).stream().filter(wl -> wl.getProductId() == productId)
+				.findFirst().orElseThrow(() -> new RuntimeException("Wishlist item not found"));
+
+		dao.delete(wishList); // Delete from DB
+	}
+}
