+++ customer-service/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=customer-wishlist
+spring.application.name=customer-service
 spring.datasource.url=jdbc:mysql://localhost:3306/microDb
 spring.datasource.username=master
 spring.datasource.password=masterkey
@@ -7,4 +7,4 @@ spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.sql.init.mode=never
 eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
-server.port = 8080
+server.port = 8083
