--- customer-wishlist/src/main/java/com/microservices/customer/wishlist/serviceImpl/ProductReviewServiceImpl.java
@@ -1,63 +1,0 @@
-package com.microservices.customer.wishlist.serviceImpl;
-
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.microservices.customer.wishlist.dao.CustomerDAO;
-import com.microservices.customer.wishlist.dao.ProductDAo;
-import com.microservices.customer.wishlist.dao.ProductReviewDAO;
-import com.microservices.customer.wishlist.entity.Customer;
-import com.microservices.customer.wishlist.entity.Product;
-import com.microservices.customer.wishlist.entity.ProductReview;
-import com.microservices.customer.wishlist.service.ProductReviewService;
-
-import jakarta.transaction.Transactional;
-
-
-@Service
-public class ProductReviewServiceImpl implements ProductReviewService {
-
-    @Autowired
-    private ProductReviewDAO productReviewDAO;
-
-    @Autowired
-    private CustomerDAO customerDAO; // Assuming you have a DAO for Customer
-
-    @Autowired
-    private ProductDAo productDAO; // Assuming you have a DAO for Product
-
-    public List<ProductReview> getReviewsByProductId(long productId) {
-        return productReviewDAO.findByProduct_Id(productId);
-    }
-
-    @Override
-    public List<ProductReview> getReviewsByCustomerId(long userId) {
-        return productReviewDAO.findByCustomer_UserId(userId);
-    }
-
-    @Override
-    public void saveProductReview(long userId, long productId, String reviewContent, int rating) {
-        ProductReview review = new ProductReview();
-        
-        // Fetch Customer and Product
-        Customer customer = customerDAO.findById(userId)
-                            .orElseThrow(() -> new RuntimeException("Customer not found"));
-        Product product = productDAO.findById(productId)
-                            .orElseThrow(() -> new RuntimeException("Product not found"));
-
-        // Set the entities
-        review.setCustomer(customer);
-        review.setProduct(product);
-        review.setReviewContent(reviewContent);
-        review.setRating(rating);
-        
-        productReviewDAO.save(review);
-    }
-
-    @Override
-    public void deleteReview(long id) {
-        productReviewDAO.deleteById(id);
-    }
-}
